Improvements & bug fixes.
Users could add & remove rows from the table. This lead to something
terrible, so this functionality has been removed.

OptionsCategoryNamesViewModel had methods to call on Page Loaded and
Unloaded events. These are now gone.

The code to register an event listener to individual categories is
replaced with use of BindingList.